+++ src/yule_orchestrator/cli/planning.py
@@ -4,8 +4,13 @@ import json
 from datetime import date, datetime
 from typing import Optional
 
-from ..planning import build_daily_plan, collect_planning_inputs, load_reminder_items, select_due_checkpoints
-from ..planning.planner import render_daily_plan
+from ..planning import (
+    build_daily_plan,
+    collect_planning_inputs,
+    load_reminder_items,
+    render_daily_plan,
+    select_due_checkpoints,
+)
 
 
 def run_planning_daily_command(
