+++ src/yule_orchestrator/planning/__init__.py
@@ -10,8 +10,10 @@ from .models import (
     PlanningTimeBlock,
     ReminderItem,
 )
-from .planner import build_daily_plan, collect_planning_inputs, load_reminder_items
-from .planner import select_due_checkpoints
+from .briefings import render_daily_plan
+from .inputs import collect_planning_inputs, load_reminder_items
+from .planner import build_daily_plan
+from .schedule import select_due_checkpoints
 
 __all__ = [
     "PlanningBlockBriefing",
@@ -27,5 +29,6 @@ __all__ = [
     "build_daily_plan",
     "collect_planning_inputs",
     "load_reminder_items",
+    "render_daily_plan",
     "select_due_checkpoints",
 ]
